fixed issues after rebase
