+++ kafka-delta-ingest/tests/helpers/mod.rs
@@ -9,14 +9,13 @@ use std::time::Duration;
 
 use bytes::Buf;
 use chrono::prelude::*;
-use deltalake::parquet::{
+use deltalake_core::kernel::{Action, Add, Metadata, Protocol, Remove, Txn};
+use deltalake_core::parquet::{
     file::reader::{FileReader, SerializedFileReader},
     record::RowAccessor,
 };
-use deltalake::protocol::{Action, Add, MetaData, Protocol, Remove, Txn};
-use deltalake::storage::DeltaObjectStore;
-use deltalake::{DeltaTable, Path};
-use kafka_delta_ingest::writer::load_object_store_from_uri;
+use deltalake_core::storage::ObjectStoreRef;
+use deltalake_core::{DeltaTable, Path};
 use kafka_delta_ingest::{start_ingest, IngestOptions};
 use rdkafka::admin::{AdminClient, AdminOptions, NewTopic, TopicReplication};
 use rdkafka::client::DefaultClientContext;
@@ -110,14 +109,13 @@ pub async fn send_bytes(producer: &FutureProducer, topic: &str, bytes: &Vec<u8>)
 // Example parquet read is taken from https://docs.rs/parquet/4.1.0/parquet/arrow/index.html#example-of-reading-parquet-file-into-arrow-record-batch
 // TODO Research whether it's possible to read parquet data from bytes but not from file
 pub async fn read_files_from_store(table: &DeltaTable) -> Vec<i32> {
-    let table_uri = table.table_uri();
-    let s3 = load_object_store_from_uri(&table_uri, None).unwrap();
-    let paths = table.get_files();
+    let s3 = table.object_store().clone();
+    let paths = table.get_files_iter();
     let tmp = format!(".test-{}.tmp", Uuid::new_v4());
     let mut list = Vec::new();
 
     for path in paths {
-        let get_result = s3.storage_backend().get(&path).await.unwrap();
+        let get_result = s3.get(&path).await.unwrap();
         let bytes = get_result.bytes().await.unwrap();
         {
             let mut file = OpenOptions::new()
@@ -126,7 +124,7 @@ pub async fn read_files_from_store(table: &DeltaTable) -> Vec<i32> {
                 .truncate(true)
                 .open(&tmp)
                 .unwrap();
-            file.write(bytes.chunk()).unwrap();
+            file.write_all(bytes.chunk()).unwrap();
             file.flush().unwrap();
             drop(file)
         }
@@ -215,6 +213,11 @@ pub async fn create_and_run_kdi(
     Runtime,
 ) {
     init_logger();
+
+    #[cfg(feature = "s3")]
+    deltalake_aws::register_handlers(None);
+    #[cfg(feature = "azure")]
+    deltalake_azure::register_handlers(None);
     println!("OPTS!: {:?}", opts);
     let topic = format!("{}-{}", app_id, Uuid::new_v4());
     let table = create_local_table(schema, delta_partitions, &topic);
@@ -403,25 +406,25 @@ pub async fn read_table_content_at_version_as<T: DeserializeOwned>(
 }
 
 pub async fn read_table_content_as_jsons(table_uri: &str) -> Vec<Value> {
-    let table = deltalake::open_table(table_uri).await.unwrap();
-    let store = load_object_store_from_uri(table_uri, None).unwrap();
+    let table = deltalake_core::open_table(table_uri).await.unwrap();
+    let store = table.object_store().clone();
     json_listify_table_content(table, store).await
 }
 
 pub async fn read_table_content_at_version_as_jsons(table_uri: &str, version: i64) -> Vec<Value> {
-    let table = deltalake::open_table_with_version(table_uri, version)
+    let table = deltalake_core::open_table_with_version(table_uri, version)
         .await
         .unwrap();
-    let store = load_object_store_from_uri(table_uri, None).unwrap();
+    let store = table.object_store().clone();
 
     json_listify_table_content(table, store).await
 }
 
-async fn json_listify_table_content(table: DeltaTable, store: DeltaObjectStore) -> Vec<Value> {
+async fn json_listify_table_content(table: DeltaTable, store: ObjectStoreRef) -> Vec<Value> {
     let tmp = format!(".test-{}.tmp", Uuid::new_v4());
     let mut list = Vec::new();
-    for file in table.get_files() {
-        let get_result = store.storage_backend().get(&file).await.unwrap();
+    for file in table.get_files_iter() {
+        let get_result = store.get(&file).await.unwrap();
         let bytes = get_result.bytes().await.unwrap();
         let mut file = File::create(&tmp).unwrap();
         file.write_all(bytes.chunk()).unwrap();
@@ -446,20 +449,16 @@ pub fn commit_file_path(table: &str, version: i64) -> String {
 }
 
 pub async fn inspect_table(path: &str) {
-    let table = deltalake::open_table(path).await.unwrap();
+    let table = deltalake_core::open_table(path).await.unwrap();
     println!("Inspecting table {}", path);
     for (k, v) in table.get_app_transaction_version().iter() {
         println!("  {}: {}", k, v);
     }
-    let store = load_object_store_from_uri(path, None).unwrap();
+    let store = table.object_store().clone();
 
     for version in 1..=table.version() {
         let log_path = format!("{}/_delta_log/{:020}.json", path, version);
-        let get_result = store
-            .storage_backend()
-            .get(&Path::parse(&log_path).unwrap())
-            .await
-            .unwrap();
+        let get_result = store.get(&Path::parse(&log_path).unwrap()).await.unwrap();
         let bytes = get_result.bytes().await.unwrap();
         let reader = BufReader::new(Cursor::new(bytes.chunk()));
 
@@ -468,15 +467,14 @@ pub async fn inspect_table(path: &str) {
         for line in reader.lines() {
             let action: Action = serde_json::from_str(line.unwrap().as_str()).unwrap();
             match action {
-                Action::txn(t) => {
+                Action::Txn(t) => {
                     println!("  Txn: {}: {}", t.app_id, t.version)
                 }
-                Action::add(a) => {
+                Action::Add(a) => {
                     let stats = a.get_stats().unwrap().unwrap();
                     println!("  Add: {}. Records: {}", &a.path, stats.num_records);
                     let full_path = format!("{}/{}", &path, &a.path);
                     let parquet_bytes = store
-                        .storage_backend()
                         .get(&Path::parse(&full_path).unwrap())
                         .await
                         .unwrap()
@@ -499,7 +497,6 @@ pub async fn inspect_table(path: &str) {
             println!("Version: {}", version);
             let log_path = format!("{}/_delta_log/{:020}.checkpoint.parquet", path, version);
             let bytes = store
-                .storage_backend()
                 .get(&Path::parse(&log_path).unwrap())
                 .await
                 .unwrap()
@@ -510,7 +507,7 @@ pub async fn inspect_table(path: &str) {
             let mut i = 0;
             for record in reader.get_row_iter(None).unwrap() {
                 let json = record.unwrap().to_json_value();
-                if let Some(m) = parse_json_field::<MetaData>(&json, "metaData") {
+                if let Some(m) = parse_json_field::<Metadata>(&json, "metaData") {
                     println!(" {}. metaData: {}", i, m.id);
                 }
                 if let Some(p) = parse_json_field::<Protocol>(&json, "protocol") {
@@ -624,7 +621,7 @@ impl TestScope {
     }
 
     pub async fn wait_on_total_offset(&self, apps: Vec<String>, offset: i32) {
-        let mut table = deltalake::open_table(&self.table).await.unwrap();
+        let mut table = deltalake_core::open_table(&self.table).await.unwrap();
         let expected_total = offset - TEST_PARTITIONS;
         loop {
             table.update().await.unwrap();
@@ -649,7 +646,7 @@ impl TestScope {
     }
 
     pub async fn validate_data(&self) {
-        let table = deltalake::open_table(&self.table).await.unwrap();
+        let table = deltalake_core::open_table(&self.table).await.unwrap();
         let result = read_files_from_store(&table).await;
         let r: Vec<i32> = (0..TEST_TOTAL_MESSAGES).collect();
         println!("Got messages {}/{}", result.len(), TEST_TOTAL_MESSAGES);
